code refactoring of ProgressAdapter classes
- eliminated TestProgressMonitor by switching to NihilistProgressAdapter for relevant tests.
- made SWTProgressAdapter a subclass of NihilistProgressAdapter to leverage common member fields and methods.
