+++ src/main/java/com/salesforce/dataloader/action/progress/NihilistProgressAdapter.java
@@ -26,9 +26,6 @@
 
 package com.salesforce.dataloader.action.progress;
 
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.LogManager;
-
 /**
  * This class implements the ILoaderProgress but does nothing with
  * the callbacks.
@@ -38,88 +35,102 @@ import org.apache.logging.log4j.LogManager;
  * @author Lexi Viripaeff
  * @since 6.0
  */
-public enum NihilistProgressAdapter implements ILoaderProgress {
-    INSTANCE;
 
-    public static NihilistProgressAdapter get() {
-        return INSTANCE;
-    }
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
-    //logger
-    private final Logger logger = LogManager.getLogger(getClass());
-    private String message = "";
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+public class NihilistProgressAdapter implements ILoaderProgress {
+    private String message;
+    private boolean success = false;
     private int numRowsWithError = 0;
+    private int numberBatchesTotal = 0;
+    private final Logger logger = LogManager.getLogger(getClass());
+    private String taskName;
+    private int workDone;
+    private int totalWork;
+    private final List<String> subTasksInOrder = new ArrayList<String>();
 
-    @Override
-    public void beginTask(String name, int totalWork) {
 
+    public NihilistProgressAdapter() {
+        // no op
     }
 
-    boolean success = false;
-    @Override
+    public void beginTask(String name, int totalWork) {
+        this.taskName = name;
+        this.totalWork = totalWork;
+    }
+    
     public void doneError(String msg) {
         success = false;
         message = msg;
         logger.error(msg);
     }
 
-    @Override
     public void doneSuccess(String msg) {
         success = true;
         message = msg;
         logger.info(msg);
-
     }
 
-    @Override
     public void worked(int worked) {
-
+        this.workDone += worked;
     }
 
     public void setTaskName(String name) {
-
+        this.taskName = name;
     }
 
-    @Override
     public void setSubTask(String name) {
+        this.subTasksInOrder.add(name);
         logger.info(name);
     }
+    
+    public String getTaskName() {
+        return this.taskName;
+    }
 
-    @Override
+    public int getTotalWork() {
+        return this.totalWork;
+    }
+    
+    public int getNumWorked() {
+        return this.workDone;
+    }
+
+    public List<String> getSubTasks() {
+        return Collections.unmodifiableList(this.subTasksInOrder);
+    }
+
     public boolean isCanceled() {
         return false;
     }
 
-    private int numberBatchesTotal = 0;
-    @Override
     public void setNumberBatchesTotal(int numberBatchesTotal) {
         this.numberBatchesTotal = numberBatchesTotal;
     }
 
-    @Override
     public boolean isSuccess() {
         return this.success;
     }
 
-    @Override
     public String getMessage() {
         return this.message;
     }
 
-    @Override
     public int getNumberBatchesTotal() {
         return this.numberBatchesTotal;
     }
 
-    @Override
     public void setNumberRowsWithError(int rowsWithError) {
         this.numRowsWithError = rowsWithError;
         
     }
 
-    @Override
     public int getNumberRowsWithError() {
         return this.numRowsWithError;
     }
-
 }
